Reorganize layer modules in Python frontend into directory
