+++ python/lbann/modules/__init__.py
@@ -1,0 +1,10 @@
+"""Neural network modules.
+
+These are a convenience for common layer patterns that are often the
+basic building blocks for larger models.
+
+"""
+
+# Import from submodules
+from lbann.modules.base import Module, FullyConnectedModule, ConvolutionModule, Convolution2dModule, Convolution3dModule
+from lbann.modules.rnn import LSTMCell, GRU
