+++ python/lbann/util/__init__.py
@@ -11,3 +11,7 @@ def make_iterable(obj):
         return obj
     else:
         return (obj,)
+
+def str_list(it, sep=' '):
+    """Convert an iterable object to a string."""
+    return sep.join(str(i) for i in make_iterable(it))
